requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
blis==0.7.4
Brotli==1.0.9
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
murmurhash==1.0.5
mysql-connector-python==8.0.25
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nbopen==0.6
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pickleshare==0.7.5
plac==1.3.3
plotly==5.0.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
Pyphen==0.10.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
smart-open==3.0.0
SQLAlchemy==1.4.19
srsly==2.4.1
tenacity==7.0.0
terminado==0.10.1
testpath==0.5.0
textstat==0.7.1
thinc==8.0.6
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
